backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
PyJWT==2.10.0
sqlparse==0.5.2
tzdata==2024.2
psycopg2-binary==2.9.10
drf-spectacular==0.28.0
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0